Split opcodes types
