+++ packages/types/src/index.ts
@@ -13,6 +13,7 @@ export * from './discord/interactions.js'
 export * from './discord/invites.js'
 export * from './discord/messages.js'
 export * from './discord/oauth2.js'
+export * from './discord/opcodes.js'
 export * from './discord/polls.js'
 export * from './discord/reference.js'
 export * from './discord/skus.js'
